Event crud (#10)
* events crud operations completed

* Added features and disabled security

+++ src/main/java/com/rescue/vscube/models/Coordinate.java
@@ -14,8 +14,9 @@ import lombok.Setter;
 @Embeddable
 @NoArgsConstructor
 public class Coordinate {
-    private float longitude;
-    private float latitude;
+
+    private double longitude;
+    private double latitude;
 
     @JsonCreator
     public Coordinate(@JsonProperty("latitude") float latitude, @JsonProperty("longitude") float longitude) {
